+++ src/test/java/com/cedarsoftware/util/io/UnknownObjectTypeTest.java
@@ -32,7 +32,7 @@ public class UnknownObjectTypeTest
     {
         String json = "{\"@type\":\"foo.bar.baz.Qux\", \"name\":\"Joe\"}";
         JsonObject myParams = TestUtil.toObjects(json, new ReadOptions().returnType(ReturnType.JSON_VALUES), null);
-        Object inputParams = JsonReader.jsonObjectsToJava(myParams, new ReadOptions());
+        Object inputParams = JsonIo.toObjects(myParams, new ReadOptions(), null);
         assert inputParams instanceof Map;
         String json2 = TestUtil.toJson(inputParams);
     }